added methods for deleting payments and events

+++ app/db_utilities.py
@@ -2,6 +2,7 @@ from app import app, db
 from hashlib import md5
 from models import User, Event, Payment
 import re
+import pdb
 
 def create_user(username, password, commit=True):
     '''Create a user with the given username and plaintext password'''
@@ -20,9 +21,17 @@ def create_user(username, password, commit=True):
     return user
 
 def remove_user(username):
+    '''
+    Remove the user with the given username
+    Return True if the user is deleted
+    '''
     user = User.query.filter(User.username == username).first()
-    db.session.delete(user)
-    db.commit()
+    if user:
+        db.session.delete(user)
+        db.commit()
+        return True
+    else:
+        return False
     
 def create_event(event_name, user_id, commit=True):
     '''Create an event with the given event name and user id'''
@@ -31,6 +40,19 @@ def create_event(event_name, user_id, commit=True):
     if commit:
         db.session.commit()
     return event
+
+def remove_event(event_id):
+    '''
+    Remove an event with the given id
+    Return True if the event is deleted
+    '''
+    event = Event.query.filter(Event.id == event_id).first()
+    if event:
+        db.session.delete(event)
+        db.session.commit()
+        return True
+    else:
+        return False
 
 def create_payment(event_id, payer, amount, commit=True):
     '''
@@ -52,3 +74,37 @@ def payment_to_float(amount):
         return float(''.join([group for group in result.groups()]))
     else:
         return float(amount)
+
+def remove_payment(id):
+    '''
+    Remove the payment with the given id
+    Return True if payment is removed
+    '''
+    payment = Payment.query.filter(Payment.id == id).first()
+    if payment:
+        db.session.delete(payment)
+        db.session.commit()
+        return True
+    else:
+        return False
+
+#utilities to check if the currently logged in owner has access to items
+def owns_event(user_id, event_id):
+    '''Return True if the user with user_id created event with event_id'''
+    event = Event.query.filter(Event.id == event_id).first()
+    if event and event.creator == user_id:
+        return True
+    else:
+        return False
+
+def owns_payment(user_id, payment_id):
+    '''
+    Returns true if the user with user_id owns the event
+    that payment_id is for
+    '''
+    payment = Payment.query.filter(Payment.id == payment_id).first()
+    if not payment:
+        return False
+    else:
+        return owns_event(user_id, payment.event_id)
+
